pyproject.toml
[project]
name = "food-orders"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.14.1",
    "coverage>=7.6.12",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pytest>=8.3.4",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
    "yarl>=1.18.3",
]
